chore: remove config class
