+++ bentoml/_internal/server/grpc/__init__.py
@@ -1,5 +1,4 @@
-from .config import Config
 from .server import Server
 from .servicer import Servicer
 
-__all__ = ["Server", "Config", "Servicer"]
+__all__ = ["Server", "Servicer"]
